refactor: Put MakeTestRequest in dedicated testing package
Utilities for testing should be separate from application code.
